기술 포럼 댓글수 표시, 기술 커뮤니티 삭제 확인 알림창

+++ multi_final/front/workspace_react/src/page/communityCommunication/CommunityDetail.js
@@ -24,10 +24,11 @@ const CommunityDetail = () => {
    navigate(`/changeCom/${no}`)
   }
   function Delete(no){
-       
-    axios.delete(`http://localhost:8085/deleteCom/${no}`)
+    if(window.confirm("게시글을 삭제하시겠습니까?")){
+      axios.delete(`http://localhost:8085/deleteCom/${no}`)
          .then(navigate('/communication')).catch(err=>console.log(err))
-      }
+    }
+  }
